backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel[asyncio]
asyncpg
python-dotenv